.internal/dr-mcp-server-lib/pyproject.toml
[project]
name = "dr-mcp-server-lib"
version = "0.1.0"
description = "DataRobot MCP Server Library - A reusable library for building MCP servers with DataRobot integration"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "datarobot-asgi-middleware>=0.2.0",
    "datarobot",
    "datarobot-predict>=1.13.2",
    "python-dotenv>=1.1.0",
    "boto3>=1.34.0",
    "httpx>=0.28.1",
    "pandas>=2.2.0",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-requests>=0.43b0",
    "opentelemetry-instrumentation-aiohttp-client>=0.43b0",
    "opentelemetry-instrumentation-httpx>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "opentelemetry-exporter-otlp-proto-http>=1.22.0",
    "aiohttp>=3.9.0",
    "aiohttp-retry>=2.8.3",
    "aiosignal>=1.3.1",
    "fastmcp>=2.12.2",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=6.2.1",
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "pydantic>=2.6.1",
    "pytest-asyncio>=0.23.5",
    "pre-commit>=3.6.0",
    "openai>=1.12.0",
    "responses>=0.25.8",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["base*"]

[tool.setuptools.package-data]
base = ["py.typed"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F", "I"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.mypy]
strict = true
python_version = "3.12"
plugins = "pydantic.mypy"
exclude = """(?x)(
    output/ # exclude files starting with "output"
    | deployment_.*/
    | include/   # or files starting with "include."
    | venv/
    | .venv/
    | .cache
    | e2e/
  )"""
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
enable_error_code = "ignore-without-code"
implicit_reexport = true
ignore_missing_imports = true
no_implicit_optional = true
show_column_numbers = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
pretty = true
show_absolute_path = true

[[tool.mypy.overrides]]
module = "datarobot.*"
implicit_reexport = true

